+++ ui/jest.config.cjs
@@ -19,7 +19,4 @@ module.exports = {
     '\\.(css|scss)$': 'identity-obj-proxy',
   },
   testEnvironment: 'jsdom',
-  env: {
-    'jest/globals': true,
-  },
 };